layout/forms/test/test_bug36619.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=36619
-->
<head>
  <title>Test for Bug 36619</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=36619">Mozilla Bug 36619</a>
<p id="display"></p>
<div id="content" style="display:none;">
  <input id='a' type='file'>
</div>
<button id='b' onclick="document.getElementById('a').click();">Show Filepicker</button>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 36619 **/

SimpleTest.waitForExplicitFinish();

var MockFilePicker = SpecialPowers.MockFilePicker;
MockFilePicker.reset();

// enable popups the first time
var oldAllow = SpecialPowers.getBoolPref("dom.disable_open_during_load");
var oldShowBar = SpecialPowers.getBoolPref("privacy.popups.showBrowserMessage");
SpecialPowers.setBoolPref("dom.disable_open_during_load", true);
SpecialPowers.setBoolPref("privacy.popups.showBrowserMessage", false);

function checkFirst()
{
  ok(MockFilePicker.shown, "File picker show method should have been called");
  document.getElementById("a").click();
  setTimeout(finishTest, 1000);
}

function finishTest()
{
  MockFilePicker.reset();
  ok(!MockFilePicker.shown, "File picker show method should not have been called");

  SpecialPowers.setBoolPref("dom.disable_open_during_load", oldAllow);
  SpecialPowers.setBoolPref("privacy.popups.showBrowserMessage", oldShowBar);
  SimpleTest.finish();
}

SimpleTest.waitForFocus(function() {
  var b = document.getElementById('b');
  b.focus(); // Be sure the element is visible.
  synthesizeMouse(b, 2, 2, {});
  SimpleTest.executeSoon(checkFirst);
});

</script>
</pre>
</body>
</html>
